requirements.txt
Django==5.0
celery==5.3.6
redis==5.0.1
ffmpeg-python==0.2.0
python-magic==0.4.27
